% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_inv_mass}
\alias{calc_inv_mass}
\title{Calculate invertebrate mass}
\usage{
calc_inv_mass(df, day)
}
\description{
Calculates invertebrate mass (grams) based on field acreage. Requires input of the fields data frame with area_ac attribute. Returns the data frame with additional columns for calculated values. Can be chained with pipes.
}
