% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_inv_mass_ts}
\alias{calc_inv_mass_ts}
\title{Calculate invertebrate mass time series}
\usage{
calc_inv_mass_ts(df, ndays)
}
\description{
A wrapper for calc_inv_mass that runs the calculation for a series of elapsed days from 1 to ndays. Returns a data frame with total invertebrate mass (grams) by field unique_id and day.
}
