% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ff_aoi}
\alias{ff_aoi}
\alias{ff_wetdry}
\title{Project Boundary}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 125 rows and 2 columns.
}
\source{
County boundaries
}
\usage{
ff_wetdry
}
\description{
This \code{sf} dataset contains a polygon defining the project boundary.
}
\examples{
head(ff_wetdry)

plot(ff_wetdry)
}
\keyword{datasets}
