% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{ff_calc_inv_mass_ts}
\alias{ff_calc_inv_mass_ts}
\title{Calculate invertebrate mass time series}
\usage{
ff_calc_inv_mass_ts(df, ndays)
}
\arguments{
\item{df}{An \code{sf} dataset, such as the provided dataframe \code{fields}, which includes an attribute area_ac containing the field area in acres. Also works with an ordinary data.frame or tibble.}

\item{ndays}{The total number of days over which to conduct the mass calculation. A sequence of days from 1 to \code{ndays} is fed into the \code{day} parameter of the \code{ff_calc_inv_mass} function.}
}
\description{
A wrapper for \code{ff_calc_inv_mass} that runs the calculation for a series of elapsed days from 1 to \code{ndays}. Returns a data frame with total invertebrate mass (kg) by field \code{unique_id} and \code{day}.
}
\examples{
# calculate the invertebrate mass time series
ff_fields |> ff_calc_inv_mass_ts(14)

# calculate the invertebrate mass time series for a particular field
ff_fields |> dplyr::filter(unique_id=="1103539") |> ff_calc_inv_mass_ts(14)
}
