% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ff_canals}
\alias{ff_canals}
\title{Canals connecting indirect return points to salmonid rearing streams}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 80 rows and 3 columns.
}
\source{
FlowWest; modified from USGS National Hydrography Dataset
}
\usage{
ff_canals
}
\description{
Supplementary `sf` geometry layer containing the non-salmonid rearing streams and canals used to calculate the flow distances from indirect return points to their nearest salmonid rearing stream (in the `ds_fbs_dist` attribute of `returns`)
}
\examples{
head(ff_canals)

plot(ff_canals$geometry)
}
\keyword{datasets}
