% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{ff_layer_fields}
\alias{ff_layer_fields}
\title{Show or hide leaflet fields layer}
\usage{
ff_layer_fields(
  m,
  show = TRUE,
  measure = "return",
  selected_return = NULL,
  selected_group = NULL,
  selected_object = NULL,
  inv_mass_days = 1
)
}
\arguments{
\item{m}{An initialized \code{leaflet} map object or \code{leafletProxy} object.}

\item{show}{A boolean value indicating whether the function call will be adding the layer to the map (\code{TRUE}) or removing the layer from the map (\code{FALSE}). Designed to be changed via \code{shiny} checkbox input by calling the function inside an observer.}

\item{measure}{A string indicating the measure to show. Choose from \code{return} to color by return type, \code{distance} to color by distance to nearest fish-bearing stream, \code{invmass} to color by modeled invertebrate mass production, or \code{wetdry} to color by wet versus dry sides.}

\item{selected_return}{The \code{return_id} of a return point to filter to, if desired.}

\item{selected_group}{The \code{group_id} of a watershed to filter to, if desired.}

\item{selected_object}{The \code{object_id} of a single field to filter to, if desired. Used internally.}

\item{inv_mass_days}{For \code{measure = "invmass"}, select the number of days to calculate the invertebrate mass production.}
}
\description{
Function to toggle the rice fields layer on an existing leaflet map, and to select which attributes to use to define the symbology of the fields.
}
\examples{
# show the layer on a leaflet map object ("m")
m <- ff_make_leaflet()
m |> ff_layer_fields(show = TRUE)

# show the layer on a leaflet map object ("m") specifying the distances measure
m |> ff_layer_fields(show = TRUE, measure = "distances")

# hide the layer on a leaflet map object ("m")
m |> ff_layer_fields(show = FALSE)

# use as part of a Shiny app with a boolean selector "show_fields" and dropdown selector "measure_fields"
if(FALSE){
  shiny::observe({
    proxy <- leaflet::leafletProxy("mainMap")
    proxy |> ff_layer_fields(show = input$show_fields,
                             measure = input$measure_fields)
  })
}

}
