% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{ff_layer_watersheds}
\alias{ff_layer_watersheds}
\title{Show or hide leaflet watersheds layer}
\usage{
ff_layer_watersheds(
  m,
  show = TRUE,
  selected_return = NULL,
  selected_group = NULL
)
}
\arguments{
\item{m}{An initialized \code{leaflet} map object or \code{leafletProxy} object.}

\item{show}{A boolean value indicating whether the function call will be adding the layer to the map (\code{TRUE}) or removing the layer from the map (\code{FALSE}). Designed to be changed via \code{shiny} checkbox input by calling the function inside an observer.}

\item{selected_return}{The \code{return_id} of a return point to filter to, if desired.}

\item{selected_group}{The \code{group_id} of a watershed to filter to, if desired.}
}
\description{
Function to toggle the watersheds layer on an existing leaflet map.
}
\examples{
# show the layer on a leaflet map object ("m")
m <- ff_make_leaflet()
m |> ff_layer_watersheds(show = TRUE)

# hide the layer on a leaflet map object ("m")
m |> ff_layer_watersheds(show = FALSE)

# use as part of a Shiny app with map "mainMap" and a boolean selector "show_canals"
if(FALSE){
  shiny::observe({
    proxy <- leaflet::leafletProxy("mainMap")
    proxy |> ff_layer_watersheds(show = input$show_watersheds)
  })
}

}
