% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{ff_make_leaflet}
\alias{ff_make_leaflet}
\title{Create leaflet map object}
\usage{
ff_make_leaflet(bbox)
}
\arguments{
\item{bbox}{A named vector defining the initial map extent via minimum and maximum lat/lon coordinates in the format \verb{c(xmin=_, ymin=_, xmax=_, ymax=_)}}
}
\description{
Initialize a leaflet map object with the desired basemap and settings.
}
\examples{
# create and display leaflet object with a specified bounding box
bbox <- sf::st_bbox(ff_fields_gcs)
m <- ff_make_leaflet(bbox)
m

# create the leaflet map for a Shiny app
output$mainMap <- leaflet::renderLeaflet({
  ff_make_leaflet(bbox)
})

}
