% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{ff_map_distances}
\alias{ff_map_distances}
\title{Interactive map of watersheds}
\usage{
ff_map_distances(selected_return)
}
\arguments{
\item{selected_return}{(optional) A specific \code{return_id} for a return point to map.}
}
\description{
Creates an interactive leaflet map showing the rice fields with watersheds return types.
}
\examples{
ff_map_distances()
ff_map_distances(selected_return = 9)
}
