% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{ff_map_invmass}
\alias{ff_map_invmass}
\title{Interactive map of invertebrate mass production}
\usage{
ff_map_invmass(day = 1, selected_return)
}
\arguments{
\item{day}{(optional) The number of days to pass to the invertebrate mass calculation. Defaults to one day if not provided.}

\item{selected_return}{(optional) A specific \code{return_id} for a return point to map.}
}
\description{
Creates an interactive leaflet map showing the rice fields with calculated invertebrate mass production.
}
\examples{
ff_map_invmass(inv_mass_days = 14)
ff_map_invmass(selected_return = 9)
}
