% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{ff_map_watersheds}
\alias{ff_map_watersheds}
\title{Interactive map of watersheds}
\usage{
ff_map_watersheds(return = NULL)
}
\arguments{
\item{return}{(optional) A specific \code{return_id} for a return point to map.}
}
\description{
Creates an interactive leaflet map showing the rice fields with watersheds return types.
}
\examples{
ff_map_watersheds()
ff_map_watersheds(return = 9)
}
