% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{ff_map_wetdry}
\alias{ff_map_wetdry}
\title{Interactive map of wet/dry sides}
\usage{
ff_map_wetdry(selected_return)
}
\arguments{
\item{selected_return}{(optional) A specific \code{return_id} for a return point to map.}
}
\description{
Creates an interactive leaflet map showing the rice fields identified by wet vs dry side.
}
\examples{
ff_map_wetdry()
ff_map_wetdry(selected_return = 9)
}
