% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fields}
\alias{fields}
\title{fields}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10106 rows and 6 columns.
}
\source{
California Department of Water Resources Land Use Program - i15 Crop Mapping YYYY; FlowWest
}
\usage{
fields
}
\description{
This `sf` dataset contains polygon geometries of rice fields based on crop mapping from YYYY.
}
\details{
- `unique_id` = the unique identifier of the rice field as defined in the original CNRA dataset
- `group_id` = the unique identifier of the watershed group, obtained via spatial join, used to join to the `watersheds` dataset
- `county` = common name of the county in which th erice field is located
- `area_ac` = area of the rice field polygon calculated in acres
- `volume_af` = inundated volume of the rice field assuming 5 inches of water
}
\keyword{datasets}
