% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishFoodMWD-package.R
\docType{package}
\name{fishFoodMWD-package}
\alias{fishFoodMWD}
\alias{fishFoodMWD-package}
\title{fishFoodMWD: Fish Food: Sacramento Valley Rice Field Prioritization for Salmonid Rearing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides access to a spatial dataset describing the Sacramento Valley rice field drainage system and potential for salmonid rearing on rice fields. Datasets include calculated drainage system characteristics influencing suitability for fish food production and/or delivery. The package also provides calculation and plotting functions and an interactive map.
}
\author{
\strong{Maintainer}: Lewis Skyler \email{slewis@flowwest.com}

Authors:
\itemize{
  \item Peng Inigo \email{ipeng@flowwest.com}
}

}
\keyword{internal}
