% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_inv_mass}
\alias{plot_inv_mass}
\title{Plot rice fields with invertebrate mass production}
\usage{
plot_inv_mass(
  day = 1,
  filename = NULL,
  width = NULL,
  height = NULL,
  units = NULL,
  colors = NULL,
  direction = 1
)
}
\arguments{
\item{day}{The day number for which to calculate the invertebrate mass. Defaults to one day.}

\item{filename}{File name to create PNG image on disk. Optional if saving the plot is desired.}

\item{width}{Plot width in \code{units}. If not supplied, uses the size of current graphics device.}

\item{height}{Plot height in \code{units}. If not supplied, uses the size of current graphics device.}

\item{units}{Units used for the width and height ("in", "cm", "mm", or "px"). Uses default \code{ggplot} dpi settings for resolution.}

\item{colors}{Color palette to use. Choose from either a character string indicating the \code{viridis} color map option to use (listed below), or a named vector defining a custom continuous gradient with \code{low} and \code{high} or a diverging gradient with \code{low}, \code{mid}, and \code{high}, e.g. \code{colors = c(low="red", mid="white", high="blue")}. For the diverging gradient, the average value is used as the midpoint.
\itemize{
\item \code{magma} (or \code{A})
\item \code{inferno} (or \code{B})
\item \code{plasma} (or \code{C})
\item \code{viridis} (or \code{D})
\item \code{cividis} (or \code{E})
\item \code{rocket} (or \code{F})
\item \code{mako} (or \code{G})
\item \code{turbo} (or \code{H})
}}

\item{direction}{If using one of the \code{viridis} scales for the \code{color} option, sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest. If -1, the order of colors is reversed.}
}
\description{
Plot all fields showing their calculated invertebrate mass production (based on acreage). The \code{calc_inv_mass} function is first run on the \code{fields} dataset using the defined number of days. Displays the watersheds, return points, and waterways on the basemap.
}
\examples{
plot_inv_mass(14)
# Choose custom colors
plot_inv_mass(14, colors=c(low="darkorange", mid="lightyellow", high="darkorchid4"))
# Returns a ggplot object that can be chained to additional ggplot functions
plot_inv_mass(14) + ggplot2::ggtitle("Rice Field Distances") + ggplot2::theme_void()
}
