% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns}
\alias{returns}
\title{returns}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 38 rows and 7 columns.
}
\source{
FlowWest
}
\usage{
returns
}
\description{
This \code{sf} dataset contains the point locations of outlets for return flow from rice field drainage networks into adjacent canals or streams.
}
\details{
\itemize{
\item \code{return_id} = the unique identifier of the return point
\item \code{return_name} = the name of the return
\item \code{return_direct} = "Direct" if directly connected to a fish-bearing stream, else "Indirect"
\item \code{ds_fbs_dist} = for Indirect returns, the distance to the downstream return to a fish-bearing stream, else zero
\item \code{ds_return_id} = for Indirect returns, the \code{return_id} of the downstream return that this return flows into, else the identifier of this same return if it directly flows into a fish-bearing stream
\item \code{ds_fbs_name} = name of the fish-bearing stream that the return drains directly or indirectly into
}
}
\keyword{datasets}
