% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riceflows4ff-package.R
\docType{package}
\name{riceflows4ff-package}
\alias{riceflows4ff}
\alias{riceflows4ff-package}
\title{riceflows4ff: Sacramento Valley Rice Field Prioritization for Salmonid Food Production}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides access to a spatial dataset describing the Sacramento Valley rice field drainage system and potential for salmonid food production on rice fields. Datasets include calculated drainage system characteristics influencing suitability for fish food production and/or delivery. The package also provides calculation and plotting functions and an interactive map.
}
\author{
\strong{Maintainer}: Skyler Lewis \email{slewis@flowwest.com}

Authors:
\itemize{
  \item Bethany Hackenjos \email{bhackenjos@gmail.com} [project director]
  \item Inigo Peng \email{ipeng@flowwest.com}
  \item Aidan Kelleher \email{akelleher@flowwest.com}
}

}
\keyword{internal}
