% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runApp}
\alias{runApp}
\title{Launch App}
\usage{
runApp()
}
\description{
\code{run_app()} launches a shiny application that allows the user to
enter data to the run fishFoodMWD functions.
}
\details{
the fishFoodMWD application can be run when the user has a dataset that they want to visualize.
}
\examples{
fishFoodMWD::run_app()
}
