% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streams}
\alias{streams}
\title{streams}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 41 rows and 3 columns.
}
\source{
Modified from CVPIA rearing habitat
}
\usage{
streams
}
\description{
Supplementary `sf` geometry layer containing the fish-bearing streams used to calculate flow distances
}
\keyword{datasets}
