% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_inv_mass}
\alias{calc_inv_mass}
\title{Calculate invertebrate mass}
\usage{
calc_inv_mass(df, day)
}
\arguments{
\item{df}{An \code{sf} dataset, such as the provided dataframe \code{fields}, which includes an attribute area_ac containing the field area in acres. Also works with an ordinary data.frame or tibble.}

\item{day}{The day number for which to calculate the invertebrate mass.}
}
\description{
Calculates invertebrate mass (grams) based on field acreage. Requires input of the fields data frame with area_ac attribute. Returns the data frame with additional columns for calculated values. Can be chained with pipes.
}
\examples{
# calculate the invertebrate mass after 14 days for each field
fields |> head(1) |> calc_inv_mass(14)
}
