% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{ff_make_leaflet}
\alias{ff_make_leaflet}
\title{Create leaflet map object}
\usage{
ff_make_leaflet(
  bbox = c(xmin = -122.3, ymin = 38.5, xmax = -121.3, ymax = 39.7)
)
}
\arguments{
\item{bbox}{A named vector defining the initial map extent via minimum and maximum lat/lon coordinates in the format \verb{c(xmin=_, ymin=_, xmax=_, ymax=_)}}
}
\description{
Initialize a leaflet map object with the desired basemap and settings.
}
\examples{
# create and display leaflet object with a specified bounding box
bbox <- sf::st_bbox(ff_fields_gcs)
ff_make_leaflet(bbox)

# create the leaflet map for a Shiny app
if(FALSE){
  output$mainMap <- leaflet::renderLeaflet({
    ff_make_leaflet(bbox)
  })
}

}
