% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ff_streams}
\alias{ff_streams}
\title{Salmonid rearing streams}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 41 rows and 3 columns.
}
\source{
Modified from CVPIA rearing habitat
}
\usage{
ff_streams
}
\description{
Supplementary `sf` geometry layer containing the salmonid rearing streams used to calculate flow distances
}
\examples{
head(ff_streams)

plot(ff_streams$geometry)
}
\keyword{datasets}
