% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishFoodMWD-package.R
\docType{package}
\name{fishFoodMWD-package}
\alias{fishFoodMWD}
\alias{fishFoodMWD-package}
\title{fishFoodMWD: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
