% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_distances}
\alias{plot_distances}
\title{Plot distances}
\usage{
plot_distances()
}
\description{
Plot all fields showing their calculated flow distances. Distance table is first joined to the fields table. Displays the watersheds, return points, and waterways on the basemap.
}
