% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fields}
\alias{plot_fields}
\title{Plot fields}
\usage{
plot_fields()
}
\description{
Plot all fields. Display the watersheds, return points, and waterways on the basemap. Returns are distinguished as draining directly into the fish-bearing stream or via a seocndary waterway.
}
