% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_watersheds}
\alias{plot_watersheds}
\title{Plot watersheds by flow type}
\usage{
plot_watersheds()
}
\description{
Plot all watersheds (groups) showing their type of flow: lateral to fish-bearing stream, direct to fish-bearing stream via return point outlet, or indirect to fish-bearing stream via return point outlet and secondary canal. Displays the watersheds, return points, and waterways on the basemap.
}
\examples{
plot_distances()
# Returns a ggplot object that can be chained to additional ggplot functions
plot_distances() + ggplot2::ggtitle("Watersheds") + ggplot2::theme_void()
}
