% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch App}
\usage{
run_app()
}
\description{
\code{run_app()} launches a shiny application that allows the user to
enter data to the run riceflows4ff functions.
}
\details{
The riceflows4ff application can be run when the user has a dataset that they want to visualize.
}
\examples{
\dontrun{
riceflows4ff::run_app()
}
}
